requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg[binary]==3.2.6

# Testing
pytest==8.2.2
